// src/app/shared/model/third-party-rslt/third-party-rslt-h-obj.model.ts
export class ThirdPartyRsltHObj {
    ThirdPartyRsltHId: number;
    SvcTypeCode: string;
    TrxTypeCode: string;
    TrxNo: string;
    CustName: string;
    IdType: string;
    IdNo: string;
    ReqDt: Date;
    ReqJson: string;
    ResultDt: Date;
    JobId: string;
    Consent: string;

    constructor() {
        this.Consent = null;
    }
}